Fully support byte-range requests
- Add MediaStreaming class that implements SteamingOutput (for readability
and reusability)
- Previous implementation ignores the "to" byte specified in Range header,
which only sends a chunk/buffer of data regardless of what the "to" value
is. This commit now handles a few more situations as below:

Range: bytes 123-
Range: bytes -123
Range: bytes -
Range: bytes 123-124

When the "from" value is not specified, it is set to 0. When the "to" byte
value is not specified, it's set to the (length - 1) of the file. When both
specified, its sends this segment as required.