Change backend hostname to be auto-detected in angular
In this project, angular app is hosted by Quarkus (as a fat jar/uber jar).
It's very inconvenient to rebuild angular app for different environment
(different hostname/ip). It is now changed to be auto-detected or that
it always points the hostname of itself, since it uses the same hostname
as the backend.