Impl WatchService to detech changes in media directory
To detect changes in media directory, the previous implementation
spawns a thread every one second to scans the whole directory. This
may be inefficient since creating new threads is expensive.

This commit uses WatchService to poll events from the file system.
Only one thread is created to constantly listening for changes. Which
may be a much better approach. However, the WatchService is initialised
during startup, which somehow cannot be properly mocked. The unit tests
are disabled temporarily.