Update methods for app lifecycle
Previously, Main#isRunning() method is provided
as a way to notify the MediaStreaming that the app
is closing. However, it doesn't seem to be working
as expected. This might be due to that the Main
class is in fact a @QuarkusMain.

An extra class is thus created (AppLifeCycle) to
catch the ShutdownEvent, and provides methods to
check if the app is still running. Hopefully, this
solution works.